# packages/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: MultiLiveTV API
  version: 1.0.0
  description: VOD aggregation API for Apple TV / iPad / Android clients
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
paths:
...
  /vod/list:
    get:
      summary: VOD list by category
      tags: [VOD]
      parameters:
        - name: sourceId
          in: query
          schema: { type: integer }
        - name: pg
          in: query
          schema: { type: integer, default: 1 }
        - name: t
          in: query
          description: Category type id (single-source)
          schema: { type: integer }
        - name: cat
          in: query
          description: Unified category slug (multi-source; takes precedence over t)
          schema: { type: string }
      responses:
        "200":
          description: OK
...
  /vod/search:
    get:
      summary: Cross-source search with merge
      tags: [VOD]
      parameters:
        - name: wd
          in: query
          required: true
          schema: { type: string }
        - name: sourceId
          in: query
          schema: { type: integer }
        - name: pg
          in: query
          schema: { type: integer, default: 1 }
      responses:
        "200":
          description: OK
...
  /vod/pic:
    get:
      summary: Poster URL for a VOD
      tags: [VOD]
      parameters:
        - name: sourceId
          in: query
          required: true
          schema: { type: integer }
        - name: ids
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
...
  /play/resolve:
    post:
      summary: Unified play resolve (direct MacCMS / ticket via bpz5)
      description: |
        Resolve a single selected play line. Clients should pick the preferred
        playable line from detail playSources (sorted by weight; skip ticket
        when ticket mode is disabled) then call this endpoint.
      tags: [Play]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mode]
              properties:
                mode:
                  type: string
                  enum: [direct, ticket]
                sourceId: { type: integer }
                url: { type: string }
                jx: { type: boolean, default: true }
                ticket: { type: string }
                providerId: { type: string }
                playFrom: { type: string }
      responses:
        "200":
          description: Resolved playable URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  url: { type: string }
                  parsed: { type: boolean }
                  mode: { type: string }
                  playFrom: { type: string }
                  expiresAt:
                    type: integer
                    nullable: true
        "501":
          description: ticket mode disabled (BPZ5_HMAC_SECRET not set)
        "502":
          description: bpz5 upstream failure
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT